# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"L5 Swagger API","description":"L5 Swagger API description","contact":{"email":"your-email@domain.com"},"version":"1.0.0"},"host":"localhost:8000","basePath":"","schemes":["http","https"],"paths":{"\/api\/news":{"get":{"summary":"News List","operationId":"NewsList","responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/api\/store":{"post":{"summary":"Add News","operationId":"AddNews","consumes":["multipart\/form-data"],"parameters":[{"name":"title","in":"formData","required":true,"type":"string"},{"name":"text","in":"formData","type":"string"},{"name":"image","in":"formData","type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/api\/show\/{id}":{"get":{"summary":"News Item","operationId":"NewsItem","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/api\/destroy\/{id}":{"get":{"summary":"News Item Destroy","operationId":"NewsItemdestroy","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/api\/update":{"post":{"summary":"News Item Update","operationId":"NewsItemUpdate","consumes":["multipart\/form-data"],"parameters":[{"name":"Item_id","in":"formData","required":true,"type":"string"},{"name":"title","in":"formData","required":true,"type":"string"},{"name":"text","in":"formData","type":"string"},{"name":"image","in":"formData","type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}},"\/api\/user-registration":{"post":{"summary":"user registration","operationId":"userRegistration","consumes":["multipart\/form-data"],"parameters":[{"name":"name","in":"formData","required":true,"type":"string"},{"name":"email","in":"formData","required":true,"type":"string"},{"name":"password","in":"formData","required":true,"type":"string"},{"name":"password_confirmation","in":"formData","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation"},"406":{"description":"not acceptable"},"500":{"description":"internal server error"}}}}},"definitions":{},"securityDefinitions":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flows":{"password":{"authorizationUrl":"\/oauth\/authorize","tokenUrl":"\/oauth\/token","refreshUrl":"\/token\/refresh","scopes":[]}}}}}